+++ src/parser/state.rs
@@ -403,6 +403,7 @@ fn state_ref_inner(input: Input<'_>) -> IResult<Input<'_>, Node<RefDecl>> {
             start,
             input,
             RefDecl {
+                extension_keywords: Vec::new(),
                 short_name: None,
                 is_derived: prefix.is_derived,
                 usage_prefix: prefix.usage_prefix,